requirements.txt
confluent-kafka
kafka-python
pandas
s3fs